Moved pfinding code to own file.
